mini_projects/keil.gyozo/AI Support Triage & Answer Drafting Agent/support-ai/requirements.txt
python >= 3.11
langchain >= 0.1.0
langgraph >= 0.0.30
openai >= 1.0.0
pydantic >= 2.0.0
python-dotenv >= 1.0.0

# Vector Database (choose one)
pinecone-client >= 3.0.0
# weaviate-client >= 4.0.0
# qdrant-client >= 1.7.0

# Re-ranking
cohere >= 4.0.0

# Optional Integrations
# zenpy  # Zendesk
# freshdesk-api
# slack-sdk
# pymsteams  # MS Teams
# jira

# Development
pytest >= 7.0.0
pytest-asyncio >= 0.21.0
black >= 23.0.0
flake8 >= 6.0.0
mypy >= 1.0.0
scikit-learn >= 1.3.0

# Logging
structlog >= 23.0.0

# FastAPI (optional)
fastapi >= 0.100.0
uvicorn >= 0.23.0